Move

Wrapper for selecting different animal movement methods.

This version uses just turn angles and step lengths to define the correlated random walk.

move(hypothesis = "crw", ...)

crw(agent, extent, stepLength, stddev, lonlat, torus = FALSE)

# S4 method for SpatialPointsDataFrame
crw(agent, extent, stepLength, stddev,
  lonlat, torus = FALSE)

# S4 method for SpatialPoints
crw(agent, extent, stepLength, stddev, lonlat,
  torus = FALSE)

Arguments

hypothesis
Character vector, length one, indicating which movement hypothesis/method to test/use. Currently defaults to 'crw' (correlated random walk) using crw.

...
arguments passed to the function in hypothesis

agent
A SpatialPoints* object. If a SpatialPointsDataFrame, 2 of the columns must be x1 and y1, indicating the previous location. If a SpatialPoints object, then x1 and y1 will be assigned randomly.

extent
An optional Extent object that will be used for torus.

stepLength
Numeric vector of length 1 or number of agents describing step length.

stddev
Numeric vector of length 1 or number of agents describing standard deviation of wrapped normal turn angles.

lonlat
Logical. If TRUE, coordinates should be in degrees. If FALSE coordinates represent planar ('Euclidean') space (e.g. units of meters)

torus
Logical. Should the movement be wrapped to the opposite side of the map, as determined by the extent argument. Default FALSE.

Value

A SpatialPointsDataFrame object with updated spatial position defined by a single occurrence of step length(s) and turn angle(s).

Details

This simple version of a correlated random walk is largely the version that was presented in Turchin 1998, but it was also used with bias modifications in McIntire, Schultz, Crone 2007.

References

Turchin, P. 1998. Quantitative analysis of movement: measuring and modeling population redistribution in animals and plants. Sinauer Associates, Sunderland, MA.

McIntire, E. J. B., C. B. Schultz, and E. E. Crone. 2007. Designing a network for butterfly habitat restoration: where individuals, populations and landscapes interact. Journal of Applied Ecology 44:725-736.

See also

pointDistance