SELES - Initiate agents

Sets the the number of agents to initiate. THIS IS NOT FULLY IMPLEMENTED.

A SELES-like function to maintain conceptual backwards compatibility with that simulation tool. This is intended to ease transitions from SELES.

You must know how to use SELES for these to be useful.

initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

# S4 method for Raster,missing,missing,ANY,missing
initiateAgents(map, numAgents, probInit, asSpatialPoints)

# S4 method for Raster,missing,Raster,ANY,missing
initiateAgents(map, probInit, asSpatialPoints)

# S4 method for Raster,numeric,missing,ANY,missing
initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

# S4 method for Raster,numeric,Raster,ANY,missing
initiateAgents(map, numAgents, probInit, asSpatialPoints)

# S4 method for Raster,missing,missing,ANY,numeric
initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

Arguments

map
RasterLayer with extent and resolution of desired return object

numAgents
numeric resulting from a call to numAgents

probInit
a Raster resulting from a probInit call

asSpatialPoints
logical. Should returned object be RasterLayer or SpatialPointsDataFrame (default)

indices
numeric. Indices of where agents should start

Value

A SpatialPointsDataFrame, with each row representing an individual agent

Examples

library(magrittr)

#> 
<<<<<<< HEAD
#> Attaching package: ‘magrittr’

#> The following object is masked from ‘package:raster’:
=======
#> Attaching package: 'magrittr'

#> The following object is masked from 'package:raster':
>>>>>>> development
#> 
#>     extract

library(raster)
library(quickPlot)

map <- raster(xmn = 0, xmx = 10, ymn = 0, ymx = 10, val = 0, res = 1)
map <- gaussMap(map, scale = 1, var = 4, speedup = 1)
pr <- probInit(map, p = (map/maxValue(map))^2)
agents <- initiateAgents(map, 100, pr)
if (interactive()) {
  clearPlot()
  Plot(map)
  Plot(agents, addTo = "map")
}
# Test that they are indeed selecting according to probabilities in pr
library(data.table)
<<<<<<< HEAD
dt1 <- data.table(table(round(map[agents], 0)))
setnames(dt1, old = "N", new = "count")
dt2 <- data.table(table(round(map[], 0)))
setnames(dt2, old = "N", new = "available")
dt <-dt1[dt2, on = "V1"]  # join the counts and available data.tables
setnames(dt, old = "V1", new = "mapValue")
dt[, selection := count/available]

#>     mapValue count available selection
#>  1:        0    NA         2        NA
#>  2:        1    NA         3        NA
#>  3:        2    NA         7        NA
#>  4:        3    11        24 0.4583333
#>  5:        4     6        16 0.3750000
#>  6:        5    29        20 1.4500000
#>  7:        6    15        13 1.1538462
#>  8:        7    17         5 3.4000000
#>  9:        8    20         9 2.2222222
#> 10:       10     2         1 2.0000000

dt[is.na(selection), selection := 0]

#>     mapValue count available selection
#>  1:        0    NA         2 0.0000000
#>  2:        1    NA         3 0.0000000
#>  3:        2    NA         7 0.0000000
#>  4:        3    11        24 0.4583333
#>  5:        4     6        16 0.3750000
#>  6:        5    29        20 1.4500000
#>  7:        6    15        13 1.1538462
#>  8:        7    17         5 3.4000000
#>  9:        8    20         9 2.2222222
#> 10:       10     2         1 2.0000000

if (interactive())
=======
dt1 <- data.table(table(round(map[agents], 0)))
setnames(dt1, old = "N", new = "count")
dt2 <- data.table(table(round(map[], 0)))
setnames(dt2, old = "N", new = "available")
dt <-dt1[dt2, on = "V1"]  # join the counts and available data.tables
setnames(dt, old = "V1", new = "mapValue")
dt[, selection := count/available]

#>    mapValue count available selection
#> 1:        0    NA         7        NA
#> 2:        1     2        17 0.1176471
#> 3:        2     5        23 0.2173913
#> 4:        3    14        19 0.7368421
#> 5:        4    21        13 1.6153846
#> 6:        5    22        11 2.0000000
#> 7:        6    32         8 4.0000000
#> 8:        7     4         2 2.0000000

dt[is.na(selection), selection := 0]

#>    mapValue count available selection
#> 1:        0    NA         7 0.0000000
#> 2:        1     2        17 0.1176471
#> 3:        2     5        23 0.2173913
#> 4:        3    14        19 0.7368421
#> 5:        4    21        13 1.6153846
#> 6:        5    22        11 2.0000000
#> 7:        6    32         8 4.0000000
#> 8:        7     4         2 2.0000000

if (interactive())
>>>>>>> development
  with(dt, {plot(mapValue, selection)})

# Note, can also produce a Raster representing agents,
# then the number of points produced can't be more than
# the number of pixels:
agentsRas <- initiateAgents(map, 30, pr, asSpatialPoints = FALSE)
if (interactive()) Plot(agentsRas)

if (require(dplyr) && getRversion() >= 3.4) {
  # Check that the agents are more often at the higher probability areas based on pr
  if (utils::packageVersion("raster") >= "2.8-11") {
    out <- data.frame(stats::na.omit(crosstab(agentsRas, map)), table(round(map[]))) %>%
             dplyr::mutate(selectionRatio = Freq / Freq.1) %>%
             dplyr::select(-layer.1, -Var1) %>%
             dplyr::rename(Present = Freq, Avail = Freq.1, Type = layer.2)
  } else {
    out <- data.frame(stats::na.omit(crosstab(agentsRas, map)), table(round(map[]))) %>%
             dplyr::mutate(selectionRatio = Freq/Freq.1) %>%
             dplyr::select(-Var1, -Var1.1) %>%
             dplyr::rename(Present = Freq, Avail = Freq.1, Type = Var2)
  }
  out
}

#> Loading required package: dplyr

#> 
#> Attaching package: 'dplyr'

#> The following object is masked from 'package:MASS':
#> 
#>     select

#> The following objects are masked from 'package:data.table':
#> 
<<<<<<< HEAD
#>     intersect, select, union

#> The following objects are masked from ‘package:stats’:
=======
#>     between, first, last

#> The following objects are masked from 'package:raster':
#> 
#>     intersect, select, union

#> The following objects are masked from 'package:stats':
>>>>>>> development
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

#>    Type Present Avail selectionRatio
#> 1     0       7     7      1.0000000
#> 2     0       0    17      0.0000000
#> 3     1      17    23      0.7391304
#> 4     1       0    19      0.0000000
#> 5     2      20    13      1.5384615
#> 6     2       3    11      0.2727273
#> 7     3      14     8      1.7500000
#> 8     3       5     2      2.5000000
#> 9     4       4     7      0.5714286
#> 10    4       9    17      0.5294118
#> 11    5       4    23      0.1739130
#> 12    5       7    19      0.3684211
#> 13    6       4    13      0.3076923
#> 14    6       4    11      0.3636364
#> 15    7       0     8      0.0000000
#> 16    7       2     2      1.0000000