SELES - Initiate agents

Sets the the number of agents to initiate. THIS IS NOT FULLY IMPLEMENTED.

A SELES-like function to maintain conceptual backwards compatibility with that simulation tool. This is intended to ease transitions from SELES.

You must know how to use SELES for these to be useful.

initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

# S4 method for Raster,missing,missing,ANY,missing
initiateAgents(map, numAgents, probInit, asSpatialPoints)

# S4 method for Raster,missing,Raster,ANY,missing
initiateAgents(map, probInit, asSpatialPoints)

# S4 method for Raster,numeric,missing,ANY,missing
initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

# S4 method for Raster,numeric,Raster,ANY,missing
initiateAgents(map, numAgents, probInit, asSpatialPoints)

# S4 method for Raster,missing,missing,ANY,numeric
initiateAgents(map, numAgents, probInit, asSpatialPoints = TRUE, indices)

Arguments

map

RasterLayer with extent and resolution of desired return object

numAgents

numeric resulting from a call to numAgents()

probInit

a Raster resulting from a probInit() call

asSpatialPoints

logical. Should returned object be RasterLayer or SpatialPointsDataFrame (default)

indices

numeric. Indices of where agents should start

Value

A SpatialPointsDataFrame, with each row representing an individual agent

Author

Eliot McIntire

Examples

if (require("sf", quietly = TRUE)) {
  library(raster)
  library(quickPlot)

  map <- raster(system.file("extdata", "map.tif", package = "SpaDES.tools"))
  names(map) <- "layer"
  pr <- probInit(map, p = (map/maxValue(map))^2)
  agents <- initiateAgents(map, 100, pr)
  if (interactive()) {
    clearPlot()
    Plot(map)
    Plot(agents, addTo = "map")
  }
  # Test that they are indeed selecting according to probabilities in pr
  library(data.table)
  dt1 <- data.table(table(round(map[agents], 0)))
  setnames(dt1, old = "N", new = "count")
  dt2 <- data.table(table(round(map[], 0)))
  setnames(dt2, old = "N", new = "available")
  dt <- dt1[dt2, on = "V1"]  # join the counts and available data.tables
  setnames(dt, old = "V1", new = "mapValue")
  dt[, selection := count/available]
  dt[is.na(selection), selection := 0]
  if (interactive())
    with(dt, {plot(mapValue, selection)})
  #'
  # Note, can also produce a Raster representing agents,
  # then the number of points produced can't be more than
  # the number of pixels:
  agentsRas <- initiateAgents(map, 30, pr, asSpatialPoints = FALSE)
  if (interactive()) Plot(agentsRas)
  #'
  if (require(dplyr)) {
    # Check that the agents are more often at the higher probability areas based on pr
    if (utils::packageVersion("raster") >= "2.8-11") {
      out <- data.frame(stats::na.omit(crosstab(agentsRas, map)), table(round(map[]))) %>%
               dplyr::mutate(selectionRatio = Freq / Freq.1) %>%
               dplyr::select(-layer.1, -Var1) %>%
               dplyr::rename(Present = Freq, Avail = Freq.1, Type = layer.2)
    } else {
      out <- data.frame(stats::na.omit(crosstab(agentsRas, map)), table(round(map[]))) %>%
               dplyr::mutate(selectionRatio = Freq/Freq.1) %>%
               dplyr::select(-Var1, -Var1.1) %>%
               dplyr::rename(Present = Freq, Avail = Freq.1, Type = Var2)
    }
    out
  }
}
#> Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 6.3.1; sf_use_s2() is TRUE
#> Loading required package: dplyr
#> 
#> Attaching package: ‘dplyr’
#> The following object is masked from ‘package:MASS’:
#> 
#>     select
#> The following objects are masked from ‘package:data.table’:
#> 
#>     between, first, last
#> The following objects are masked from ‘package:raster’:
#> 
#>     intersect, select, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#>    Type Present Avail selectionRatio
#> 1     0       1     1     1.00000000
#> 2     0       0     3     0.00000000
#> 3     2       3     4     0.75000000
#> 4     2       0     5     0.00000000
#> 5     3       2    14     0.14285714
#> 6     3       2    14     0.14285714
#> 7     4       5    20     0.25000000
#> 8     4       0    19     0.00000000
#> 9     5      12    12     1.00000000
#> 10    5       2     6     0.33333333
#> 11    6      12     1    12.00000000
#> 12    6       2     1     2.00000000
#> 13    7      15     1    15.00000000
#> 14    7       5     3     1.66666667
#> 15    8       7     4     1.75000000
#> 16    8      12     5     2.40000000
#> 17    9       7    14     0.50000000
#> 18    9       5    14     0.35714286
#> 19   10       4    20     0.20000000
#> 20   10       2    19     0.10526316
#> 21   11       1    12     0.08333333
#> 22   11       0     6     0.00000000
#> 23   12       1     1     1.00000000
#> 24   12       0     1     0.00000000